requirements.txt
streamlit
openai
python-dotenv
pydantic
